+++ bls12_377/src/curves/g1.rs
@@ -1,14 +1,22 @@
-use ark_ec::models::{
-    short_weierstrass::{Affine as SWAffine, SWCurveConfig},
-    twisted_edwards::{
-        Affine as TEAffine, MontCurveConfig, Projective as TEProjective, TECurveConfig,
+use ark_ec::{
+    bls12,
+    hashing::curve_maps::wb::{IsogenyMap, WBParams},
+    models::{
+        short_weierstrass::{Affine as SWAffine, SWCurveConfig},
+        twisted_edwards::{
+            Affine as TEAffine, MontCurveConfig, Projective as TEProjective, TECurveConfig,
+        },
     },
     CurveConfig,
 };
 use ark_ff::{Field, MontFp, Zero};
-use core::ops::Neg;
+use ark_std::ops::Neg;
 
+use super::g1_swu_iso::{SwuIsoParameters, ISOGENY_MAP_TO_G1};
 use crate::{Fq, Fr};
+
+pub type G1Affine = bls12::G1Affine<crate::Parameters>;
+pub type G1Projective = bls12::G1Projective<crate::Parameters>;
 
 #[derive(Clone, Default, PartialEq, Eq)]
 pub struct Parameters;
@@ -157,6 +165,12 @@ pub const G1_GENERATOR_X: Fq = MontFp!("8193799937315096423993825557346594823998
 /// G1_GENERATOR_Y =
 /// 241266749859715473739788878240585681733927191168601896383759122102112907357779751001206799952863815012735208165030
 pub const G1_GENERATOR_Y: Fq = MontFp!("241266749859715473739788878240585681733927191168601896383759122102112907357779751001206799952863815012735208165030");
+
+impl WBParams for Parameters {
+    type IsogenousCurve = SwuIsoParameters;
+
+    const ISOGENY_MAP: IsogenyMap<'static, Self::IsogenousCurve, Self> = ISOGENY_MAP_TO_G1;
+}
 
 // The generator for twisted Edward form is the same SW generator converted into
 // the normalized TE form (TE2).