+++ bls12_377/src/curves/g2.rs
@@ -1,12 +1,19 @@
 use ark_ec::{
-    models::{short_weierstrass::SWCurveConfig, CurveConfig},
-    short_weierstrass::Affine,
+    bls12,
+    hashing::curve_maps::wb::{IsogenyMap, WBParams},
+    models::CurveConfig,
+    short_weierstrass::SWCurveConfig,
 };
+
 use ark_ff::{Field, MontFp, Zero};
 
 use crate::{g1, Fq, Fq2, Fr};
 
-pub type G2Affine = Affine<Parameters>;
+use super::g2_swu_iso::{SwuIsoParameters, ISOGENY_MAP_TO_G2};
+
+pub type G2Affine = bls12::G2Affine<crate::Parameters>;
+pub type G2Projective = bls12::G2Projective<crate::Parameters>;
+
 #[derive(Clone, Default, PartialEq, Eq)]
 pub struct Parameters;
 
@@ -76,3 +83,9 @@ pub const G2_GENERATOR_Y_C0: Fq = MontFp!("6316029476829207320938136194393519890
 /// G2_GENERATOR_Y_C1 =
 /// 149157405641012693445398062341192467754805999074082136895788947234480009303640899064710353187729182149407503257491
 pub const G2_GENERATOR_Y_C1: Fq = MontFp!("149157405641012693445398062341192467754805999074082136895788947234480009303640899064710353187729182149407503257491");
+
+impl WBParams for Parameters {
+    type IsogenousCurve = SwuIsoParameters;
+
+    const ISOGENY_MAP: IsogenyMap<'static, Self::IsogenousCurve, Self> = ISOGENY_MAP_TO_G2;
+}